Update gitignore and refactor config schema
