--- configs/2-be-updated/blood_rage_ages.json
@@ -1,88 +1,0 @@
-{
-    "metadata": {
-        "name": "Blood Rage",
-        "assets_dir": "blood_rage"
-    },
-    "phases": [
-        {
-            "name": "Age 1",
-            "imgs": "assets/blood_rage/gods_gift/images",
-            "audio": "assets/blood_rage/gods_gift/audio"
-        },
-        {
-            "name": "Age 2",
-            "imgs": "assets/blood_rage/action/images",
-            "audio": "assets/blood_rage/action/audio"
-        },
-        {
-            "name": "Age 3",
-            "imgs": "assets/blood_rage/ragnarok/images",
-            "audio": "assets/blood_rage/ragnarok/audio"
-        }
-    ],
-    "endings": [
-        {
-            "key": "K_r",
-            "name": "Raven",
-            "img": "assets/blood_rage/raven/images",
-            "audio": "assets/blood_rage/dunk.mp3"
-        },
-        {
-            "key": "K_b",
-            "name": "Bear",
-            "img": "assets/blood_rage/bear/images",
-            "audio": "assets/blood_rage/dunk.mp3"
-        },
-        {
-            "key": "K_s",
-            "name": "Serpent",
-            "img": "assets/blood_rage/serpent/images",
-            "audio": "assets/blood_rage/dunk.mp3"
-        },
-        {
-            "key": "K_w",
-            "name": "Wolf",
-            "img": "assets/blood_rage/wolf/images",
-            "audio": "assets/blood_rage/dunk.mp3"
-        },
-        {
-            "key": "K_m",
-            "name": "Ram",
-            "img": "assets/blood_rage/ram/images",
-            "audio": "assets/blood_rage/dunk.mp3"
-        }
-    ],
-    "sfx": [
-        {
-            "name": "Power Up",
-            "key": "K_1",
-            "audio": "assets/sfx/power_up.mp3"
-        },
-        {
-            "name": "Horn",
-            "key": "K_2",
-            "audio": "assets/sfx/horn.mp3"
-        },
-        {
-            "name": "Roar",
-            "key": "K_3",
-            "audio": "assets/sfx/roar.mp3"
-        },
-        {
-            "name": "Angel",
-            "key": "K_4",
-            "audio": "assets/sfx/angel.mp3"
-        },
-        {
-            "name": "Bell",
-            "key": "K_5",
-            "audio": "assets/sfx/bell.mp3"
-        },
-        {
-            "name": "Crow",
-            "key": "K_6",
-            "audio": "assets/sfx/crow.mp3"
-        }
-    ],
-    "font": "assets/blood_rage/fonts/Norse.otf"
-}
